Muzzle doesn’t seem to like those helper classes there
Had to move them externally for muzzle to be happy.
